FUNCIONA?
pofavo

+++ NetBeansProjects/PapucplanetCineDA/src/pe/edu/pucp/papucplanet/cine/mysql/FuncionMySQL.java
@@ -129,7 +129,7 @@ public class FuncionMySQL implements FuncionDAO{
                 funcion.getPelicula().setIdPelicula(rs.getInt("fid_pelicula"));
                 funcion.getPelicula().setTitulo(rs.getString("titulo"));
                 funcion.getPelicula().setGenero(Genero.valueOf(rs.getString("genero")));
-                funcion.getPelicula().setDuracion(rs.getDouble("duracion"));
+                funcion.getPelicula().setDuracion(rs.getInt("duracion"));
                 funcion.getPelicula().setSinopsis(rs.getString("sinopsis"));
                 funcion.getPelicula().setImagenPromocional(rs.getString("imagen_link"));
                 
@@ -170,7 +170,7 @@ public class FuncionMySQL implements FuncionDAO{
                 funcion.getPelicula().setIdPelicula(rs.getInt("fid_pelicula"));
                 funcion.getPelicula().setTitulo(rs.getString("titulo"));
                 funcion.getPelicula().setGenero(Genero.valueOf(rs.getString("genero")));
-                funcion.getPelicula().setDuracion(rs.getDouble("duracion"));
+                funcion.getPelicula().setDuracion(rs.getInt("duracion"));
                 funcion.getPelicula().setSinopsis(rs.getString("sinopsis"));
                 funcion.getPelicula().setImagenPromocional(rs.getString("imagen_link"));
                 
@@ -216,7 +216,7 @@ public class FuncionMySQL implements FuncionDAO{
                 funcion.getPelicula().setIdPelicula(rs.getInt("fid_pelicula"));
                 funcion.getPelicula().setTitulo(rs.getString("titulo"));
                 funcion.getPelicula().setGenero(Genero.valueOf(rs.getString("genero")));
-                funcion.getPelicula().setDuracion(rs.getDouble("duracion"));
+                funcion.getPelicula().setDuracion(rs.getInt("duracion"));
                 funcion.getPelicula().setSinopsis(rs.getString("sinopsis"));
                 funcion.getPelicula().setImagenPromocional(rs.getString("imagen_link"));
                 
@@ -266,7 +266,7 @@ public class FuncionMySQL implements FuncionDAO{
                 func.getPelicula().setIdPelicula(rs.getInt("fid_pelicula"));
                 func.getPelicula().setTitulo(rs.getString("titulo"));
                 func.getPelicula().setGenero(Genero.valueOf(rs.getString("genero")));
-                func.getPelicula().setDuracion(rs.getDouble("duracion"));
+                func.getPelicula().setDuracion(rs.getInt("duracion"));
                 func.getPelicula().setSinopsis(rs.getString("sinopsis"));
                 func.getPelicula().setImagenPromocional(rs.getString("imagen_link"));
                 
@@ -400,7 +400,7 @@ public class FuncionMySQL implements FuncionDAO{
                 Pelicula pelicula = new Pelicula();
                 pelicula.setIdPelicula(rs.getInt("id_pelicula"));
                 pelicula.setTitulo(rs.getString("titulo"));
-                pelicula.setDuracion(rs.getDouble("duracion"));
+                pelicula.setDuracion(rs.getInt("duracion"));
                 pelicula.setGenero(Genero.valueOf(rs.getString("genero")));
                 pelicula.setSinopsis(rs.getString("sinopsis"));
                 pelicula.setImagenPromocional(rs.getString("imagen_link"));
